PycharmProjects/prompt-eng/requirements.txt
nicegui
google-generativeai
python-dotenv